Separated most of the non async methods out into their own files, DiceRoll is dicerolling code, WebParser is for printing lookups, added lines to .gitignore
